# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Multi-Institution Admission and Student Onboarding Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents for an application
      parameters:
        - name: applicationId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Uploaded document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}/verify:
    post:
      operationId: verifyDocument
      tags: [documents]
      summary: Verify a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentVerification"
      responses:
        "200":
          description: Verified document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  # ─── INVENTORY ────────────────────────────────────────────────────────────────
...
  /dormitories:
    get:
      operationId: listDormitories
      tags: [dormitories]
      summary: List dormitories for institution
      parameters:
        - name: institutionId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dormitories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dormitory"
    post:
      operationId: createDormitory
      tags: [dormitories]
      summary: Create dormitory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DormitoryInput"
      responses:
        "201":
          description: Dormitory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dormitory"
  /dormitories/{id}:
    patch:
      operationId: updateDormitory
      tags: [dormitories]
      summary: Update dormitory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DormitoryUpdate"
      responses:
        "200":
          description: Dormitory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dormitory"

  # ─── NOTIFICATIONS ────────────────────────────────────────────────────────────
...
  /dashboard/institution/{id}:
    get:
      operationId: getInstitutionDashboard
      tags: [dashboard]
      summary: Institution dashboard stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Institution stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstitutionStats"
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Recent activity feed
      parameters:
        - name: institutionId
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /dashboard/application-pipeline:
    get:
      operationId: getApplicationPipeline
      tags: [dashboard]
      summary: Application counts by status stage
      parameters:
        - name: institutionId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Pipeline
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PipelineStage"
  /dashboard/revenue-summary:
    get:
      operationId: getRevenueSummary
      tags: [dashboard]
      summary: Revenue summary by month
      parameters:
        - name: institutionId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Revenue by month
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueMonth"

  # ─── AUDIT LOGS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Document:
      type: object
      required: [id, applicationId, documentType, status, createdAt]
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        documentType:
          type: string
          enum: [admission_letter, birth_certificate, result_slip, passport_photo, medical_form, leaving_certificate, transfer_letter, nemis]
        fileUrl:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: ["string", "null"]
        verifiedBy:
          type: ["integer", "null"]
        createdAt:
          type: string
    DocumentInput:
      type: object
      required: [applicationId, documentType, fileUrl]
      properties:
        applicationId:
          type: integer
        documentType:
          type: string
        fileUrl:
          type: string
    DocumentVerification:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]
        rejectionReason:
          type: string

    # ─── INVENTORY ────────────────────────────────────────────────────────────
...
    Dormitory:
      type: object
      required: [id, institutionId, name, capacity, occupancy]
      properties:
        id:
          type: integer
        institutionId:
          type: integer
        name:
          type: string
        gender:
          type: ["string", "null"]
          enum: [male, female, mixed, null]
        capacity:
          type: integer
        occupancy:
          type: integer
        dormitoryMaster:
          type: ["string", "null"]
        dormitoryMasterPhone:
          type: ["string", "null"]
        occupancyPercent:
          type: ["number", "null"]
    DormitoryInput:
      type: object
      required: [institutionId, name, capacity]
      properties:
        institutionId:
          type: integer
        name:
          type: string
        gender:
          type: string
        capacity:
          type: integer
        dormitoryMaster:
          type: string
        dormitoryMasterPhone:
          type: string
    DormitoryUpdate:
      type: object
      properties:
        name:
          type: string
        gender:
          type: string
        capacity:
          type: integer
        dormitoryMaster:
          type: string
        dormitoryMasterPhone:
          type: string

    # ─── NOTIFICATIONS ────────────────────────────────────────────────────────
...
    InstitutionStats:
      type: object
      required: [totalApplications, pendingDocuments, reportingStudents, totalRevenue, inventoryAlerts, dormitoryOccupancy]
      properties:
        totalApplications:
          type: integer
        approvedApplications:
          type: integer
        pendingDocuments:
          type: integer
        reportingStudents:
          type: integer
        totalRevenue:
          type: number
        pendingPayments:
          type: integer
        inventoryAlerts:
          type: integer
        dormitoryOccupancy:
          type: number
        checkedInToday:
          type: integer
    ActivityItem:
      type: object
      required: [id, type, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        studentName:
          type: ["string", "null"]
        institutionName:
          type: ["string", "null"]
        createdAt:
          type: string
    PipelineStage:
      type: object
      required: [status, count]
      properties:
        status:
          type: string
        count:
          type: integer
        label:
          type: string
    RevenueMonth:
      type: object
      required: [month, revenue]
      properties:
        month:
          type: string
        revenue:
          type: number
        payments:
          type: integer
...